Wire up the hotlels list.
Previous to this commit, scrolling the hotels list was causing no transition at all. Now we do our transitions,
entering states, etc. The transitions look really bad and in somecases backwards at this point, but the key here
is that things are starting to get wired up, and we can polish the animations and things all in good time.