Set active listeners based on state.
This does two things, it makes sure our non-active controller doesnt start firing rouge events and hosing us.
It also and more importantly prevents cross states. E.g. We set ResultsState.HOTELS, and we tell the flight
controller this, and it reports back that it is at ResultsFlightsState.LIST_DOWN, which the activity interprets
as overview mode, but we just said we were in Hotles mode, so overview mode is wrong. Now we just dont listen to
anything the flights controller says when we are in hotels mode.