Tablet flight details UI/UX fix
Move the statistics header into the ScrollView such that it
scrolls along with the flight line content. All the content
in the white section of flight details is meant to scroll.

Also, give this container a white background with rounded
corners at the bottom to match UI polish from Chris.