Fix bug that prevents us from touching anything in flights mode ever.
As the changelog shows, we are setting setTouchPassThroughEnabled(false) if our hotels state is not ROOMS_AND_RATES,
but that is failing to consider the HOTEL_LIST_DOWN state (which is overview) and since the hotels controller sits on top
of the flights contorller, and the map fragment is left VISIBLE with alpha(0) so that it renders in the background,
this was blocking all touches to the flights controller. We update the code to also allow touchpassthrough when we are in
HOTEL_LIST_DOWN mode.