2335 - money w/ no decimal places bug fix
The previous implementation of formatRate had logic that had the cached
NumberFormat depend on currencyCode, flags, and BigDecimal amount in order
to format money properly, despite caching the NumberFormat on just currencyCode
and flags.

This commit refactors the logic to check the amount.scale() special case and
encode this data into the flags. This means we can use our existing caching
logic and it will work.