2377 - Fix AccountManager multiple account bug.
So somehow someone got in a crazy state, and since we never wanted people to get into that state, we threw
a RuntimeException. Ultimately we dont want the app crashing, so we removed that exception and instead
do an AccountManager clear before adding our newly logged in user. This should prevent us from having multiple
users fairly well, and if we do ever get into that crazy state again (which should be super difficult) nothing
should break.