2372 Fix flight line airport code clipping bug.
This was a nightmare to track down. The root cause is 2 fold. First all of our flight line code is dependant
on waypoint.getBestSearchTime(). This make sense because it was developed for flight search, but is a bummer
because we also use it for itins where we should be using getMostRelevantTime(). This issue I did not fix,
because it would be a bigger change than this small bug merits.

The second issue was this. The itin where this bug was found had an intermediate flight that landed AFTER
the final flight was supposed to land. So a flight got rerouted,diverted,whatever, but we do our line position
calculations based on where our flight starts/ends in relation to where the first flight starts and the
last flight ends. This intermediate flight ended after when the last flight in the itin was supposed to land
and so it drew itself in the right spot, but unfortunately that right spot was off the end of the view and thus
got clipped. This 2 line change was an entire day in the making. Fuuuuuuuuuuuuuuuuuuuuuuuuuuuuuuuuuuuuuuuu.