Tablet Results lists refactor.
This commit replaces FruitScrollUpListview with FruitList - the primary difference here being that FruitList
is percentage based, and it implements IStateProvider. There were a few issues with the previous implementation,
the worst of which was the inability to position the list when it wasnt being drawn. Now we can set percentage
at any time, and when it draws it should draw in the right position.

We also updated the structure of the ResultsListFragment/ResultsFlightListFragment/ResultsHoelListFragment so that
the base class ResultsListFragment now does more of the work and we have less duplicate code.

This commit isnt totally complete, the FruitList class still needs some custom attributes, and needs to handle overscrolling better
along with fixing some strange flininging bugs. Still, things are mostly working now, and this commit is already way too large.