FlightTripView refactor pt 2
Add another layer of abstraction for how the view is drawn. This
adds a little more structure, such that the lines and circles aren't
entirely dependent on flight, and its relative time. The relative
time algorithm should not be changed. This commit sets the stage
for having a more flexible circle/line view be drawn, something
that is more or less hardcoded.