2411 - Fix google wallet not showing for flights bug.
This bug originated in the rewrite of the flights checkout animation. In the past the FlightCheckoutFragment would
not be attached until the user was entering checkout mode (which is after the trip call is complete), now it gets
attached right away, so that it exists and can be smoothly slid into place. However, at the time of FlightCheckoutFragment.onCreate
being called, our trip call is typically not finished, so before this change, it would say that the trip didnt support
google wallet. After the trip call if you rotated, it would check correctly and show the google wallet button appropriately.

Now we wait until after we have a valid trip (the trip has an itinerary number) to determine if the trip supports GW.