2315 - Ensure we use appropriate email for flight bookings.
In flights checkout we send two parameters to the api, &email and &mainFlightPassenger.email, unbeknownst to us,
&mainFlightPassenger.email takes precedence. Previously we were for sure putting the correct email in the &email
param, but sometimes we were sending a different email in the &mainFlightPassenger.email param. This would typically
be an email address manually entered with the the Traveler on the hotels side. If the user were to select that same
traveler on the flights side, its email address would be used. Meaning the confirmation page would display the wrong
email address, and hypothetically the user would not know where their confirmation got sent. Pretty lame, but fixed now.