2413 - Fix flight checkout back press gui glitch.
This was an odd bug, sometimes hitting back when you were in checkout mode on flights would not scroll all the way
to the top. This fix was sort of a lucky guess. Our smooth scroll was happening, and then our pricebar fragment
would get attached and I think some sort of layout happened at that point causing the scroll to stop. Now we just
wait for the fragment transaction to finish before we begin smooth scrolling and it seems to solve it.
(cherry picked from commit dbaaa6b09b4c833a321a5000b9149b3595aa432f)