2352 Hotel confirmation crash on share (from v3.2)
We were clearing out the HotelSearch data in the SearchResultsFragmentActivity.onDestroy()
(which was being invoked due to ActivityKillReceiver). This meant the selectedProperty
would get cleared from Db shortly after the confirmation page loaded, and thus would go boom
when attempting to share.

We've opted to stuff the Property into the BookingResponse to prevent the crash and keep
the code working as close as possible to how it was beforehand.