2423 - Fix GoogleWallet hotel line items bug.
Before this fix it was possible for the api to return values that did not always add up to total. When this happened
we would create google wallet line items that wouldnt add up to total price, and as a result google wallet would fail.

Now if the line items dont add up to total, we just fudge the least significant line item. So the sum of line items
always matches total. Hopefully one day api will put out a fix for this, but whatcha gonna do.