Fix ResultsFlightMapFragment rendering all the time bug.
So I think someone misunderstood the OnGlobalLayoutListener - from the docs:
Callback method to be invoked when the global layout state or the visibility of views within the view tree changes

The view tree contains all of the views, so any time there was a visibility change, this thing would redraw itself, but
more than that, it would cause a draw loop on the map. We fix that by caching some values and not drawing again unless they
change.