2376 change airline filter behavior
Before we started with every airline checked, and if you wanted to single
out an airline, you would have to un-check a bunch of airlines. Now we've
decided to start the airlines as un-checked, and when you starat checking
airlines, only flights from those airlines will show up. If no airlines
are checked, the entire result set is shown.