2393. Fixed issue where flights were added to tripBucket even when user had just seen the details.
We need to keep track of when the user has actually added the flight,
as opposed to just checking to see if the user has selected it.
This works fine on the phone due to single pane flow but on tablet we need to save/check if the user has selected.

auditors: brad