2447 flight selection back and forth bug fix
The root of the problem is that in order to book a successful trip, you must select
a valid combination of flight legs that have been given to use in the flight search
response. Before any client-side filtering occurs, we filter out invalid trips behind
the scences, based upon which flight legs you have already selected. We correctly
filter out invalid trips from the inbound leg as you select different outbound legs.
One problem is that we weren't clearing out the selected inbound leg when going back
to the outbound leg, and thus intereacting with the filters would cause the flight
results list to essentially disappear in some cases.

Additionally, we needed to have a notion of a "selected" FlightTrip and an "added" FlightTrip.
The added FlightTrip is used for trip bucket. The selected FlightTrip is used in the flights
selection path.