2444 CheckBox and its text padding problem
API 17 changed the behavior of the padding* attributes in CheckBox
to refer to padding relative to the TextView. This means I can
successfully set a padding between the CheckBox and its text
via paddingLeft attribute. For API < 17, we need a little bit more
padding (e.g. the width the CheckBox drawable + text padding).