2441 stops filter copy change
Plurals are limited in that the "zero" entry only gets selected if the given
locale has special grammar rules for the quantity zero. English and most languages
do not, therefore our special-case "No Stops" label was not being used. Now we
hardcode in the Java code to use the correct string based on quantity.