2441 stops filter - update filter pane to reflect actual flight trips
We are better about updating the inbound legs filter pane to reflect the
actualy filtering that can take place after the inbound result set is
diminished (due solely to selecting the outbound leg). The problem is,
in some cases, we could pretty much just get rid of the entire content
of the filter pane, if say, there's only one flight in the inbound
flight list that will match the outbound leg. We'll have to figure out
exactly what we want to do here.