2427. (Phone) Actual move to checkoutV2 is happening.
I did make this all one big commit on purpose, since i thought it all ties up to make the checkout happen.
And would be a lot easier for someone to see the whole flow change. Then again I could have probably broken it down.

What's changed with new Checkout(v2):
We have to make the "create" API call before we call the new checkout.
AND everytime before we add add coupons, as well as after we remove the added coupon.

I haven't handled error cases for the "create" call. Will be making a seperate commit for that soon.