Added dummy Traveler forms fragment. Improved expanding checkout item animation.
So the traveler form still needs some thought. The idea is going to be that it will generate a traveler entry form based on LOB. So
we just tell it we are booking a hotel, and it exposes the appropriate fields. This should be strait forward enough, but what parts
should be shared with the payment entry form if any. Also, the mocks have a total lack of detail on how the Redress # entry thing will look
and there is absoluately no way to select a stored traveler the way the mocks have things now. We are taking baby steps and getting a feel
for the cases that we will be handling and more or less trying some things out to see how they will work.