1883: Fix flight map line cutout
A BitmapShader is positioned around the origin of the window coordinates
so we just have to translate it underneath where we want it to draw.
Having this work at all was very lucky, I'm surprised I didn't notice it
was completely broken sooner.

I also adjusted the map padding to fix the arrival pin on the map all
the time.