2466 tablet search calendar bug fix
When setting the state of the CalendarPicker externally (i.e. by its
controller, rather than by having the user interact with the calendar),
we want to make sure the month gets drawn correctly. So now when the start
and end dates get set, we also set the month to be the month of the start date.