2492. Fixed Coupon-related crash picking new room
Since we now use the new Db.getHotelSearch().isCouponApplied() instead of relying on the Db.getHotelSearch().getCreateTripResponse() != null
to check if a coupon is applied. Let's reset the isCouponApplied() to false when we exit the activity.

NOTE: We can't rely on Db.getHotelSearch().getCreateTripResponse() != null to check because now we have to call the new "create" API
when we add/remove a coupon which is passed the CreateTripResponseHandler already.