Add support for interpolators in StateManager.
Slowing down and adding a bounce interpolator to all of the results animations is pretty funny.
I tried out a bunch, and the default seems to look best in most cases. Still want the ability to
use a custom interpolator for StateManager transitions though.

An interesting thought might be to have different animation styles in different parts of the app.
E.g. Flights decel and hotels accel or something. I dont know, just wanted to liven up the animations a bit.