2494 tablet search data persistence update
So, in order to have the Results and Checkout launchers be useful we had
previously been relying Db.saveOrLoadForTesting() which would essentially
save/load the previous HotelSearch/FlightSearch to and from disk.

Eventually we will not be supporting the Results and Checkout launchers and
will want to have a more complete disk persistence strategy. This commit
attempts to match the hotel/search disk data with what is in memory by saving
the hotel/flight search whenever a new search happens and also when things like
the "selected" (or "added") item changes.

Additionally, because we are still actively developing this app we don't want the
hotel search data to be cleared from disk every 30 minutes, so I bypass the timeout
such that the data should stick around indefinitely. Once we get closer to launch
we'll want to remove that.

The next steps are to implement the network calls on checkout such the checkout Activity
recovers itself more properly (and can book rooms and flights and stuff).