Fix crasher on SectionBillingInfo.
This was checking Db.getFlightSearch.getSelectedFlightTrip().isCardTypeSupported(...) and if
there was no value we would NPE. It was a case that shouldnt be happening, but it is causing
pain for myself as developer, so it is now fixed.

Also, the AS formatter had some things to say about this file.