Add proper dev setting support for Hotels suppress final booking
Hotels checkout domain v2 supports the same suppressFinalBooking flag
that flights has had on the final check out call. I've switched the dev
setting to reflect this such that you can suppress a flight and or suppress
a hotel, separately. This gets rid of the "spoof" hotel bookings option
which would not actually make the check out call and just simulate a
successful response.

Also, I've updated the automation code to reflect the changes.