Add TripBuck disk persistence to Db
Writing to data to disk and loading that data back up is a common
occurrence in Db.java. I've attempted to standardize these operations
be creating some interfaces that can be implemented to perform the
actual read/write operation.

The benefit of this is that it is easier to write persistence code
for new data that pops up in Db, because error handling and stats
logging is taken care of for free.

This commit introduces the disk persistence utilities and also an
example of how it is used, by using them for TripBucket save/load.
In the future all of the disk read/write methods will be re-written
to use the utility methods, but I want this code to go through a little
review before then. It seems like the utilities could be generalized
even more to operate solely on an array of java.lang.Field, which
could potentially be pretty slick.