2488: Clip Paths that are too large
If a `Path` is too large (larger than the screen size) then we need to
clip it or the framework won't be able to rasterize it and upload it to
the GPU.

Even still there appears to be a bug in the output path from the
`Region`. We need to do a transformation with the identity to get it to
show up at all. A little extra work, but hey it renders.