On confirmation when we show the book button, dont make it push another activity, just update the old one.
This is still a little bogus, because the Controller needs to update which fragments it shows. So currently
if we book a hotel and then click the book flights button, we end up on the overview screen, but only the hotel
data entry forms will be showing, and other similar stuff.