Store FlightSearchState instead of FlightSearch in TripBucketItemFlight
We were running into issues with storing the whole FlightSearch object
in TripBucketItemFlight because it was the same FlightSearch as the
one stored in Db and also because it is very large. It makes more sense
to store the FlightSearchState as this juncture, although, ideally the
TripBucket stores more state - enough state to keep track of a flight
from a totally different FlightSearch than the one we are keeping in Db.