2526 - show hotel stay total in trip bucket
We always want to show the total price the user will pay for their hotel
checkout in the trip bucket card. If the card is collapsed, we show the
price rounded. When expanded, the price is not rounded. This mirrors what
we do for flights and also makes sure the user sees the total price on the
results screen in the trip bucket after they've added a hotel to the trip bucket.