2486: Fix bad math
So very bad. I needed to move the shift I computed into the
`preTranslate` of the Viewport Matrix. Now things center on the screen
correctly.

This really points out that we need to add some fidelity to the map.
Having the great lakes will make it much easier to distinguish what is
going on when flying from middle part of the US.