Tablet Results Loading - Jam the actual search download call into the ResultsActivity.
So this is not what we ultimately want. The individual controller fragments should be in charge of loading their own data. Currently the
ExpediaServicesFragment is built around being able to use retainInstance == true, which throws an exception if we try to attach it as a
child fragment. This will be addressed in time.