Move data loading into the controller fragments.
So everything about this should be ok, except we removed a call to TripBucket load from disk, and this is because
we are in sort of a sticky situation now with our searches coming back at different times. We might need to start
saving the tripbucket in separate parts so we can load trip bucket items based on LOB. Our trip bucket should mostly
be in memory, the only case this is maybe bad in is when we were background killed and want to resume. I dont know
I need to think about it some more.