2576. Fixed Cannot Get to Hotel Checkout.
This is a temporary urgent fix, so that QA isn't blocked.
This doesn't fix the rotate issue, which is something that will be done after i'm done with the refactor.2576. Fixed Cannot Get to Hotel Checkout.

This is a temporary urgent fix, so that QA isn't blocked.
This doesn't fix the rotate issue, which is something that will be done after i'm done with the refactor.