GDE flight histogram view on results screen
This commit stuffs the histogram view in to the results screen. The designs
aren't finalized but this is a basic prototype of how it could work. Basically,
we want it to work like this for now:

1. When updating search paramaters, kick off a gde flight search. This comes back
quickly and we can populate a view of the histogram for departure dates and prices
in the flights pane on the results overview screen.

2. We also kick off a much longer full flight search when search params change. You
should see a progress indicator in the histogram pane when a flight search is in progress.

3. Click on the "60 flights" on the flight list or "When To Fly" TextView to switch between
the histogram and flight list. This only works if a flight search is sitting in memory.

Lots of stuff left to do:

1. Make the histogram interactable.

2. Better histogram UI.

3. lots of bug fixes probably