TabResRef - Fix/simplify add to bucket animations.
Sorry for the big commit. Basically we got rid of the IAddToTripListener implementations, and try to rely on states more.
The things no longer fly towards the bucket, but since this animation and the new trip bucket are not really well defined
I wanted to pull out the complicated broken logic, and start with a clean slate when we have a better idea of how the new
animations are going to work. This looks decent for now. Atleast the correct elements are in play.