2522.Checkout V2 error handling - Hotels (Part 1)
We can unify the error handling for both flights and hotels.
This is the first part of the refactoring which readies things to be used and unified in the next commit.
Also (important) this takes care of the SESSION_TIMEOUT error for hotels, by showing an appropriate pop
and asking the user to choose another hotel again. (Previously it was pointing to flights error dialog - Fixed)