TabResRef - StateListenerCollection improvements.
Use HashSet instead of arraylist for inactive listeners - should get a small perf improvement.

More importantly, this solves the following issue: Say one of our fragments has child fragments and it chooses
to remove one of them in onStateFinalized. This is totally acceptable behavior, except if the fragment we are
removing happens to be an IStateListener, it tries to remove itself while we are iterating over the Listener list
and then we end up throwing a concurrent modification exception. Now we just wait to change the lists until
we are done iterating. Note also that we are keying on the listener itself, so if we try to perform two actions
on the same listener, the most recent will take precedence.