2553: Cache OkHttpClient
There is no need to create a new one on every request. The OkHttpClient
even synchronizes access to the cookie store so we don't have to be
careful about synchronization there.

However we cannot use the cached client for requests that need to ignore
cookies. We create an adhoc one in those instances.