2553: Much simpler cookie store
I was making it way too complicated. We just want to send all the
cookies all the time. Our `CookiePolicy` is going to make sure it comes
from the server we expect it to on release. On debug thought we need to
send all cookies so things like the mock proxy work.