Flight histogram statistical analysis to remove outliers
The prices sometime seem to contain outliers, e.g. most flights
returned are in the 300-500 range, but there's one flight with a
price of $1300. This not only bones our relative graph, but it also
means we are showing the user a shitty deal.

This commit introduces some simple quartile analysis inspired by
old math classes where I was forced to draw box and whisker plots.

It uses a simple heuristic to remove outliers. I think we will
probably want to tweak the heuristic as I see it removing prices
that don't really seem like outliers.