Flight search histogram UI - add colors, for science
So, I was thinking of ways to better visualize this flight
data we get back from GDE, and I decided to assign the price
a color based upon which quartile range the price falls under.
Green means this deal is cheap, red means this deal is on the
expensive end.

My main idea is to have these colors on the calendar, as
somewhat of a calendar heatmap, pointing to you the best days
to select...but that's not gonna happen unless we're sure we
want to do that...