Add new waypoint picker to tablet launch.
This commit is pretty big and messy and incomplete. So we now move towards using the same waypoint chooser
for the different tablet screens. We remove some functionality like current location search, because it
ultimatly should act the same way it does on results, and that functionality needs to be pulled out.