Results - Rework the downloads a little bit.
Prior to this commit the download frags were set up to begin downloading whenever they were attached.
This was ok, but it made it a little odd to restart a download inline. Further we now are more careful
about when we clear things in db, so hopefully we will run into fewer NPEs because the frags that depend on
Db should nolonger be active when we clear our Db values.