TabFlightsRef - Begin refactoring the way the flights flow works in EBad tablet.
Ultimately this refactor will result in support for N legs.
It also aims to simplify the TabletResultsFlightsController by taking over much of the leg
selection logic and animations.
Futhermore it creates much more sane States. It is not sustainable to have states like
FLIGHT_TWO_FILTERS as we do not want to have to add states like  FLIGHT_THREE_FILTERS in
the future, which would result in a whole bunch of new code to support animations etc.