Fix tablet checkout blurred bg image not showing up
Now that we request specific image sizes for the background destination images
on the tablet results and checkout screens, we want the bitmaps to be the same
dimensions in each case. Because the background image's dimensions are reliant
on the container it will be shown (e.g. entire content background), we need to
match the content size for results/checkout activities.

This commit changes TabletCheckoutActivity to ActionBar overlay mode so the
background images are the same dimension (e.g. to match the previous Activity
behaviors).