2667. Refactoring FlightBookingFragment.
This will be a repeat of what we did for HotelBookingFragment, where in
all the network related calls are handled by their respective fragments so it can be shared b/w tablet and phone.

This commit takes care of tablet only.
Now TripBucketFlightFragment is only responsible for the views and is stripped off making n/w calls.