L2ImageCache - cache instances refactor
I was never really happy with how DestinationImageCache was set up as a singleton
instance of L2ImageCache. The architecture felt weird as the class was called
one thing, but "getInstance()" returned an instance of a different class (L2ImageCache).
In addition, access just felt clunky.

I've moved all the cache instances as public static members inside of L2ImageCache. This means
L2ImageCache is somewhat overloaded, but I think that it makes sense in this case. It is easier
to find the cache you're looking for (e.g. just go to L2ImageCache).

Also, I marked the cache instance as public without a getter. In my opinion:

	L2ImageCache.sDestination > L2ImageCache.getDestinationImageCache()

I don't see the benefit of verbosity in this case.