Attempt to smooth out some animations.
So if the framerate is slow, sometimes the last frame to go through will be like 0.87, and so then our
ui sits at 0.87 percent of the way through the transition until a bunch of gui work gets done on finalize
and this causes a pretty serious little jump at the end of a bunch of our animations. This fix basically
just says if you are near the end of an animation, just consider it the end and draw accordingly.