Fix a visual bug in choosing flight legs where an empty space would show above the first flight.
This is odd. So I think what is happening is that the row dividers in the list view decide not to draw when the item above them is no longer visible,
so we end up with this odd behavior where it looks like our list is filling in spaces correctly until our last row goes above the fold, and then we
show whatever is behind the list until we scroll up a few more pixels. I solved this in the cheapest way possible, by setting the background of the list.

This behavior does not cause problems on the first leg, so something odd is going on, but for now this fix will do.