Fix StateListenerCollection bug w.r.t rotation.
So the StateListenerCollection takes a default state. If we call registerStateListener with fireFinalize=true
typically we notify the listener of the last state, and if none was available we would provide this default state.

Well if we rotate and restore a state on rotation, but the listenercollection was declared during the construction
of the instance, then we end up having the wrong state in here, and we tell listeners about it. The new behavior is to
just do nothing if we don't yet have a state, because if we don't yet have an explicit state, we should probably expect
someone to set one sometime very soon.