Fix ConcurrentModificationExceptions bug.
In the past I had tried solve this problem with all this pre/post iterate non-sense. After googling a little
it looks like ArrayList.clone() uses an array level copy and should be really quick. Further we never expect
there to be too many listeners... I think right now we probably top out at about 5, so the this shouldnt cause
a huge performance hit, and it is safer and simpler. Win win win.