Add new mode to FrameLayoutTouchController - prevent mashing.
So sometimes the user just wants to mash the screen, but if we react to every single
touch event our app cant keep up and we end up in a strange state or exploding. This
new mode allows for a touch cooldown period, in which no new touch events will be passed
to children until we have not received any new touches for the cooldown duration.