Add GDE Flight histogram back into the results screen.
Sorry for the mega-commit.
We first updated the ResultsFlightHistogramFragment to be a simple ListFragment.
Then we created the new ResultsGdeFlightsFragment to house it.
The new ResultsGdeFlightsFragment handles downloading and displaying GDE results.
We added the ResultsGdeFlightsFragment to the SearchController.

Eventually the ResultsGdeFlightsFragment will be able to control search params.
At that time it will have mechanisms for showing loading state and switching
between single leg and round trip gde response stuff.