Label non-infant child travelers by age.
Matching the initially searched for ages to Traveler objects
is a bit more involved than I would like, but is a necessary
step.

We must indicate the age of the child whose info the user is
editting because the age must correspond to the Traveler's
passengerCategory. If the traveler's birthdate does not match
the passengerCategory, then the checkout will fail. There
has been discussion around indicating age once the user is
actually on the traveler activity, but this makes it testable
for now.