Tablet flight details - baggage fee information updates
Rather than popping up a WebViewActivity for the payment fees, this commit adds
the WebViewFragment to ResultsFlightDetailsFragment as a child fragment. This allows
for displaying the baggage link in the flight details card pane, and allows for animation.

This commit necessitated updating the FlightUtils.configureBaggage utility method to support
different functionality for phone and tablet versions.