2723: Converted amenity icons to SVG
This is really much simpler than it appears. I removed all the
ic_amenity_*.png files from everywhere, and added res/raw/ic_amenity_*.svg.
Then I modified the snippet_amenity.xml to take a SVGView
and modified LayoutUtils.addAmenities to handle SVGs instead of pngs.
The amenity icon colors are different between phone and tablet, so a
ColorFilter is used to colorize the (originally black) amenity SVGs.
That's it!