This is the first of the tablet dateless search commits.
When we land on the results screen, we kick off a search for tonight for hotels.
We display the calendar and gde results immediatly, and dont do any flight searches
until the user chooses a date.

We are still missing a bunch of stuff, such as updating the calendar when we click on gde
,being able to clear the search dates, better gde stuff in general, better loading states,
better error messages etc...