Add a param update cooldown to the flight/hotel controllers.
The idea here is that when the user drags the calendar, we dont actually need to start
and cancel a new search for every date they drag over, it makes more sense to wait a
little while for the user input to slow down and then do our work.

This seems to reduce the number of crashes on dragging the calendar around, which was
the intention. It should also save us some money, because flight searches are expensive.