Tablet results stability refactor.
In this commit we attempt to solve a number of bugs associated with lifecycle on tablet
results. We introduced the IAcceptingListenersListener which solves the following problem:
When a parent fragment wants to listen to one of its child fragments, the lifecycle of
the child is more important than the parent because it will pause before the parent and
resume afterwards, meaning the parent doesnt really know when is best to be listening to
its children. Now we use the IAcceptingListenersListener interface so that the children
can simply tell the parents when they are ready to be listened to or not based on lifecycle.

So far this seems to have been quite effective. Future work will probably include moving
our SearchController frags to implementing IAcceptingListenersListener and to get our
download frags to act more predictably based on lifecycle.