WebViewFragment refactor (also fixes 2749)
We don't want scroll to persist after loading new URLs. I couldn't find a reliable way
to scroll the WebView back to the top after loading a new URL, so I just construct a
new WebView instance when loading a different url (via bind).

This commit re-organizes WebViewFragment code such that it is more readable and flexible,
and also updates the bind method to call the correct code in order to load a new URL
after initial construction.