Speed up transition to HotelDetails.
This animation was running between 5 and 15 fps on a 2013 N7, which is unacceptable. This commit
brings us up to about 25-35 fps on the same device. We are still running un-acceptably slow on
the N10, and we'd like to get closer to 60 fps all around. Note that a number of timing loggers
are still around, and that is to help with future optimizations.

The improvements:
* Use a view and alpha to dim the hotel map. This way we use a hardware layer and never need to call draw.
* Delay view updates until they are needed. The hotel controller was doing work before it needed to.
* Remove global layout listener in the hotel details fragment. This is where we made the biggest gains.
  Prior to this commit, we were listening for global layout changes, but because this frag was invisible until
  it was transitioning in, it was forcing a layout in mid animation which totally killed our framerate. Now
  we measure using the measurement listener, and this way we set layout before we are doing our transition.

Note: It looks like the hotel details arrow that points to the map is not positioned correctly any more, and
will be fixed in a subsequent commit.