2787 tablet flight re-selection bug fix
It was not working 100% correctly to select a new flight after already adding
one to the trip bucket. We make sure to reset the the whole query after selecting
a flight for the trip bucket. Also, we make sure to refer to the TripBucket data
store in the TripBucketFragments so crashes don't ensue when we reset the "selection"
data stored in the query.