Add rough sketch of infant toggle to sorts and filters
This should only come up when there is at least one infant in
flight search params. If a user decides they want their infant(s)
to instead be in their own seats or vice versa, toggling this
spinner will (eventually) kick off a new search...all from the
comfort of flights mode.

It looks ugly, and it is ugly. BUT, beneath the grit and grime
is a foundation for a better child traveler experience.