Fighting the OOM fight
This bumps down the mem cache size for the destination image cache down
from 2 images to 1 for the tablet flow. We only need 1 full screen destination
image in memory at a given time anyway, so the extra Bitmap is waste and
we need to cut down on all the waste we can!