Tablet checkout crash fix - child travelers
This fixes that crash where a null LocalDate gets passed to a utility method
and boom (ReadableInstant can not be null or w/e).  It looks like basically
there was an assumption that the Traveler birthdate would be set in Flights
flow, but this is not always the case because hotels flow can also mess with
the Travelers in Db.