Rip out Chris Banes code. Improve color averaging performance greatly.
The Chris Banes color scheme algorithm was nice, and it generating nice colors from images, but it was not performant.
Its memory footprint, where N is the number of pixels, was O(5N/2) which is not great. Even worse it was computationally intensive.
The Chris Banes code would take between 250ms and 1000ms on a typical hotel/destination image. The new code uses O(N + 1) memory and
typically runs in < 1ms.

Note: We have removed all the darkening we were doing before. This will be re-added once design gets back to me with hsv adjustment values.