2780 child traveler search experience on tablet
So, child travelers are kind of complicated, and it gets even more complicated
when you're dealing with infants (children under two). If you're flying with
an infant you have two options: (1) have the infant sit in your lap or (2) purchase
a seat for the infant.

We don't want to overload the guest picker on search, especially because it is
shared between hotels and flights, and hotels flow doesn't care about infant
in lap vs. infant in seat. Moreover, all infants must abide by the same rule (e.g.
you can't book one infant in lap and one infant in seat for the same booking).

Given all the above information, we default flight searches with infants to be "in lap".
We think this is a safe assumption, and we afford the user the opportunity to change
this in the flight filters, where, we kick them off a new search if they would like
to change their infant seating preference. This commit pops a dialog confirming the
changed preference and kicks off a new search with the correct parameters.